refactor: remove bc7 flavor and billingclient dimension
Drop Billing Client 7 support and the entire bc7/bc8 flavor dimension,
keeping only the (former bc8) Billing Client 8 behavior as the single
build. No more -bc7 artifacts will be published.

- Promote purchases/src/bc8 sources into src/main and delete src/bc7
- Remove the billingclient dimension and bc7/bc8 flavors from all modules
- Rename catalog keys bc8 -> billingClient / billing-bc8 -> billing
- Collapse Bc7/Bc8 variant + task names across CI, Fastlane and scripts
- Delete the redundant api-defaults-bc7.txt signature
