+++ .run/Debug tests.run.xml
@@ -1,5 +1,5 @@
 <component name="ProjectRunConfigurationManager">
-  <configuration default="false" name="BC8 Debug tests" type="GradleRunConfiguration" factoryName="Gradle">
+  <configuration default="false" name="Debug tests" type="GradleRunConfiguration" factoryName="Gradle">
     <ExternalSystemSettings>
       <option name="executionName" />
       <option name="externalProjectPath" value="$PROJECT_DIR$" />
@@ -10,7 +10,7 @@
       </option>
       <option name="taskNames">
         <list>
-          <option value="testDefaultsBc8DebugUnitTest" />
+          <option value="testDefaultsDebugUnitTest" />
         </list>
       </option>
       <option name="vmOptions" />
