+++ examples/admob-sample/build.gradle.kts
@@ -41,9 +41,8 @@ android {
         adUnitOverride("ADMOB_REWARDED_INTERSTITIAL_AD_UNIT_ID")
 
         // Library modules have dimensions used to separate different APIs and billing client versions.
-        // Applications don't need these, so we default to the "defaults" and "bc8" flavors.
+        // Applications don't need this, so we default to the "defaults" flavor.
         missingDimensionStrategy("apis", "defaults")
-        missingDimensionStrategy("billingclient", "bc8")
 
         vectorDrawables {
             useSupportLibrary = true
