+++ examples/paywall-tester/build.gradle.kts
@@ -30,18 +30,6 @@ android {
 
         missingDimensionStrategy("apis", "defaults")
 
-        flavorDimensions += "billingclient"
-
-        productFlavors {
-            create("bc8") {
-                dimension = "billingclient"
-                isDefault = true
-            }
-            create("bc7") {
-                dimension = "billingclient"
-            }
-        }
-
         vectorDrawables {
             useSupportLibrary = true
         }
