+++ examples/rcttester/build.gradle.kts
@@ -19,18 +19,6 @@ android {
         // Library modules have a dimension used to separate different APIs.
         // Applications don't need this, so we default to the "defaults" flavor.
         missingDimensionStrategy("apis", "defaults")
-
-        flavorDimensions += "billingclient"
-
-        productFlavors {
-            create("bc8") {
-                dimension = "billingclient"
-                isDefault = true
-            }
-            create("bc7") {
-                dimension = "billingclient"
-            }
-        }
     }
 
     signingConfigs {
