+++ feature/amazon/build.gradle.kts
@@ -5,19 +5,6 @@ plugins {
 android {
     namespace = "com.revenuecat.purchases.amazon"
 
-    // billingclient dimension is added for bc7/bc8 support
-    flavorDimensions += "billingclient"
-
-    productFlavors {
-        create("bc8") {
-            dimension = "billingclient"
-            isDefault = true
-        }
-        create("bc7") {
-            dimension = "billingclient"
-        }
-    }
-
     defaultConfig {
         missingDimensionStrategy("apis", "defaults")
     }
