+++ feature/galaxy/build.gradle.kts
@@ -11,19 +11,6 @@ metalava {
 android {
     namespace = "com.revenuecat.purchases.galaxy"
 
-    // billingclient dimension is added for bc7/bc8 support
-    flavorDimensions += "billingclient"
-
-    productFlavors {
-        create("bc8") {
-            dimension = "billingclient"
-            isDefault = true
-        }
-        create("bc7") {
-            dimension = "billingclient"
-        }
-    }
-
     defaultConfig {
         missingDimensionStrategy("apis", "defaults")
     }
