+++ gradle.properties
@@ -50,7 +50,7 @@ rcTesterVersionName=1.0
 
 #Do not sign releases. When calling uploadArchives pass -PRELEASE_SIGNING_ENABLED=true
 RELEASE_SIGNING_ENABLED=false
-ANDROID_VARIANT_TO_PUBLISH=defaultsBc8Release
+ANDROID_VARIANT_TO_PUBLISH=defaultsRelease
 SONATYPE_HOST=CENTRAL_PORTAL
 SONATYPE_AUTOMATIC_RELEASE=true
