--- purchases/src/bc7/kotlin/com/revenuecat/purchases/google/BillingResultExtensionsBillingIndependent.kt
@@ -1,7 +1,0 @@
-package com.revenuecat.purchases.google
-
-import com.android.billingclient.api.BillingResult
-
-internal fun BillingResult.toHumanReadableDescription() =
-    "DebugMessage: $debugMessage. " +
-        "ErrorCode: ${responseCode.getBillingResponseCodeName()}."
