+++ scripts/api-dump.sh
@@ -2,8 +2,7 @@
 
 exit_code=0
 
-./gradlew metalavaGenerateSignatureDefaultsBc8Release || exit_code=$?
-./gradlew metalavaGenerateSignatureCustomEntitlementComputationBc8Release || exit_code=$?
-./gradlew metalavaGenerateSignatureDefaultsBc7Release || exit_code=$?
+./gradlew metalavaGenerateSignatureDefaultsRelease || exit_code=$?
+./gradlew metalavaGenerateSignatureCustomEntitlementComputationRelease || exit_code=$?
 
 exit $exit_code
