+++ test-apps/e2etests/build.gradle.kts
@@ -15,7 +15,6 @@ android {
         versionCode = 1
         versionName = "1.0"
         missingDimensionStrategy("apis", "defaults")
-        missingDimensionStrategy("billingclient", "bc8")
     }
 
     buildTypes {
