+++ test-apps/testpurchasesandroidcompatibility/build.gradle.kts
@@ -15,7 +15,6 @@ android {
         versionName = "1.0"
 
         missingDimensionStrategy("apis", "defaults")
-        missingDimensionStrategy("billingclient", "bc8")
 
         testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
         multiDexEnabled = true
